web/backend/requirements.txt
# Web Framework
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-multipart

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary

# AI/ML - CPU compatible
easyocr==1.7.0
numpy==1.24.3
opencv-python-headless==4.11.0.86

# Authentication & Security
passlib==1.7.4
PyJWT==2.10.1

# Image Processing
Pillow==11.3.0

# API Integration
openai==1.107.1
Requests==2.32.5

# Email
fastapi_mail==1.5.0

# Utils
python-dotenv==1.1.1
pydantic==2.11.7

# Stripe
stripe==10.0.0